added mosquitto test fixture
This adds a fixture which spawns a real mosquitto for the duration of
the test. This takes a couple of seconds to get rolling because the
spawn needs to actually get the server up. There is also a potential
race around port allocation, however, it's pretty insignificant.

If mosquitto can't be started, tests are skipped.