update rain roll over algorithm
This will take any rain that registers from the last record of the
day, to the first record of the next day, and put it in the next day
bucket. This will prevent weird artifacts on Weather Underground where
the total for yesterday comes in as the first packet of today.

This carry forward logic is pretty heavily tested.

It also reduces the amount of traffic on the MQTT bus, as we're only
doing the compute totals during a rollover event.